+++ src/main/java/com/logmate/streaming/messaging/producer/impl/kafka/KafkaDlqProducer.java
@@ -1,12 +1,17 @@
-package com.logmate.streaming.producer.impl.kafka;
+package com.logmate.streaming.messaging.producer.impl.kafka;
 
-import com.logmate.streaming.common.constant.kafka.KafkaConstant;
-import com.logmate.streaming.producer.DlqProducer;
+import com.logmate.streaming.global.constant.kafka.KafkaConstant;
+import com.logmate.streaming.messaging.producer.DlqProducer;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 
+/**
+ * Kafka DLQ(Dead Letter Queue) Producer
+ *
+ * Kafka 전송 실패 또는 처리 불가능한 로그를 DLQ 토픽으로 전달하는 책임을 가진 프로듀서.
+ */
 @Slf4j
 @Service
 @RequiredArgsConstructor
@@ -17,8 +22,10 @@ public class KafkaDlqProducer implements DlqProducer {
 
   @Override
   public void send(String key, String message, Throwable cause) {
-    log.warn("[KafkaDlqProducer] Sending message to DLQ. key={}, cause={}",
-        key != null ? key : "null", cause != null ? cause.getMessage() : "unknown");
+    log.warn("[KafkaDlqProducer] Sending message to DLQ. topic={}, key={}, cause={}",
+        kafkaConstant.topic.DLQ_TOPIC,
+        key != null ? key : "null",
+        cause != null ? cause.getMessage() : "unknown");
 
     kafkaTemplate.send(kafkaConstant.topic.DLQ_TOPIC, key, message)
         .whenComplete((result, ex) -> {
@@ -29,7 +36,11 @@ public class KafkaDlqProducer implements DlqProducer {
                 result.getRecordMetadata().partition(),
                 result.getRecordMetadata().offset());
           } else {
-            log.error("[KafkaDlqProducer] Failed to send message to DLQ", ex);
+            log.error("[KafkaDlqProducer] Failed to send message to DLQ. topic={}, key={}, error={}",
+                kafkaConstant.topic.DLQ_TOPIC,
+                key != null ? key : "null",
+                ex.getMessage(),
+                ex);
           }
         });
   }
